requirements.txt
openai==2.6.0
Pillow==12.0.0
protobuf==6.33.0
psycopg2_binary==2.9.11
python-dotenv==1.1.1
python-telegram-bot==22.5
PyYAML==6.0.3
Requests==2.32.5
ruamel.base==1.0.0
